test(coins/padAddress): move tests to coins-only file so symlinked defi test stays clean
